fix(rasterize): close budget, NaN, cull-direction and dispatch gaps
Four fixes from the branch's final review:

- choose_tiling's pair-budget check summed n_tiles in int32, which
  mx.sum does not promote: for a mesh whose true (tile, face) pair
  count reaches 2**31, the reduction silently wraps. A positive wrap
  reports a pair count under budget while the true count is billions,
  so build_tile_lists allocates an undersized keys buffer that the
  scatter kernel then writes past with no bound check -- an unbounded
  out-of-bounds GPU write. Reduce in int64 before comparing against
  MAX_PAIRS (and widen the offsets cumsum for the same reason).

- A vertex with w == 0 in clip space gives x/w = NaN, which rounds and
  casts to fixed-point 0 -- inside every bound, so the existing
  Inf/saturation guard never sees it. NaN comparisons are false, so
  the kernel's depth test doesn't reject it either: a bogus triangle
  gets rasterized with a garbage face index and no error. Check
  finiteness on the projected x/y before the cast. (While in this
  file, also drops the now-dead `pixel_center_fx` -- exported but
  referenced only by its own tests; the kernel hardcodes the same
  arithmetic and cannot call Python anyway.)

- Every existing cull test asserted only that "back" and "front"
  differ, which a coordinated sign flip across the binning code and
  the oracle would still pass. Document which winding "back" removes
  (counter-clockwise is front-facing, OpenGL convention -- verified
  against the shipped mapping, which does not flip y) and pin the
  direction with a concrete test.

- The spec's failure-modes table asks for background arrays with no
  kernel dispatch when every face is degenerate, culled, or offscreen
  (total_pairs == 0), not only when num_faces == 0. Add the missing
  short-circuit.

Confirmed the pair-budget and NaN regression tests fail with
"DID NOT RAISE" against the pre-fix code, and confirmed the cull
direction test fails when the binning code's back/front sign is
flipped (the exact defeat scenario the old XOR-shaped test could not
catch).

+++ src/mlx_arsenal/rasterize/_binning.py
@@ -158,7 +158,13 @@ def choose_tiling(
     last = 0
     for ts in candidates:
         bounds, n = face_spans(verts_fx, faces, width, height, ts, cull)
-        total = item_int(mx.sum(n))
+        # n is int32 and mx.sum does not promote: for a mesh whose true pair
+        # count reaches 2**31, an int32 reduction silently wraps (verified:
+        # 4097 screen-spanning faces at 16384px, tile 16 -> true total
+        # 4,296,015,872 reduces to 1,048,576, which is under MAX_PAIRS). That
+        # defeats the budget check below and lets an unbounded scatter write
+        # through in build_tile_lists. Widen before reducing.
+        total = item_int(mx.sum(n.astype(mx.int64)))
         if tile_size is not None or total <= MAX_PAIRS:
             return ts, bounds, n, total
         last = total
@@ -246,7 +252,10 @@ def build_tile_lists(
             mx.zeros((num_tiles + 1,), dtype=mx.int32),
         )
 
-    offsets = (mx.cumsum(n_tiles, axis=0) - n_tiles).astype(mx.int32)
+    # Safe in int32 once the budget check above is honest (total_pairs <=
+    # MAX_PAIRS = 32M), but widen anyway so this cumsum carries no dependency
+    # on that reasoning holding elsewhere.
+    offsets = (mx.cumsum(n_tiles.astype(mx.int64), axis=0) - n_tiles).astype(mx.int32)
     params = mx.array([num_faces, tiles_x], dtype=mx.int32)
 
     (keys,) = _get_scatter_kernel()(
